system tray: split up code
Move the system tray item and data box into the system_tray module and move the platform specific code into ksni.rs and tray_icon.rs as those are going to grow.
